BroadAddress.py v0.0
The original BroadAddress.py file. This *may* be improved in the future to give you the first and last usable IP Address within a Class-C network. I may also construct a sort of graphical table to clean up the look of the results it gives you in the end. It also needs to check the user's input, and I'll have to construct a loop so that the program restarts and lets you enter a new IP Address and Subnet Mask.

Use this as you wish, and even modify it. Just give me the credit for the original file if you do.